+++ tempest_fastapi_sdk/__init__.py
@@ -342,7 +342,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.78.0"
+__version__: str = "0.79.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",